test: refactoring test case
