+++ engine/src/test/java/com/arcadedb/ACIDTransactionTest.java
@@ -550,70 +550,6 @@ public class ACIDTransactionTest extends TestHelper {
     });
   }
 
-  @Test
-  public void testExplicitLock() {
-    final Database db = database;
-
-    final int TOT = 100;
-
-    database.getSchema().getOrCreateVertexType("Node");
-
-    final AtomicInteger committed = new AtomicInteger(0);
-    final AtomicInteger caughtExceptions = new AtomicInteger(0);
-
-    final RID[] rid = new RID[1];
-
-    database.transaction(() -> {
-      final MutableVertex v = db.newVertex("Node");
-      v.set("id", 0);
-      v.set("name", "Exception(al)");
-      v.set("surname", "Test");
-      v.save();
-      rid[0] = v.getIdentity();
-    });
-
-    final int CONCURRENT_THREADS = 4;
-
-    // SPAWN ALL THE THREADS AND INCREMENT ONE BY ONE THE ID OF THE VERTEX
-    final Thread[] threads = new Thread[CONCURRENT_THREADS];
-    for (int i = 0; i < CONCURRENT_THREADS; i++) {
-      threads[i] = new Thread(() -> {
-        for (int k = 0; k < TOT; ++k) {
-          try {
-            database.transaction(() -> {
-              database.acquireLock().type("Node").lock();
-
-              MutableVertex v = rid[0].asVertex().modify();
-              v.set("id", v.getInteger("id") + 1);
-              v.save();
-            });
-            committed.incrementAndGet();
-
-          } catch (Exception e) {
-            caughtExceptions.incrementAndGet();
-          }
-        }
-
-      });
-      threads[i].start();
-    }
-
-    // WAIT FOR ALL THE THREADS
-    for (int i = 0; i < CONCURRENT_THREADS; i++)
-      try {
-        threads[i].join();
-      } catch (InterruptedException e) {
-        // IGNORE IT
-      }
-
-    assertThat(database.countType("Node", true)).isEqualTo(1);
-
-    assertThat(rid[0].asVertex().getInteger("id")).isEqualTo(CONCURRENT_THREADS * TOT);
-    assertThat(committed.get()).isEqualTo(CONCURRENT_THREADS * TOT);
-    assertThat(caughtExceptions.get()).isEqualTo( 0);
-    assertThat(committed.get() + caughtExceptions.get()).isEqualTo(TOT * CONCURRENT_THREADS);
-  }
-
   private void verifyDatabaseWasNotClosedProperly() {
     final AtomicBoolean dbNotClosedCaught = new AtomicBoolean(false);
